requirements.txt
streamlit
openai-whisper
huggingface-hub
python-dotenv
gradio
